Remove deprecated metods
