+++ Auth0/DesktopWebAuth.swift
@@ -1,17 +1,13 @@
 #if os(macOS)
-import Cocoa
+import AppKit
 import AuthenticationServices
 
-/**
- Resumes the current Auth session (if any).
+extension NSApplication {
 
- - parameter urls: urls received by the macOS application in AppDelegate
- - warning: deprecated as the SDK will not support macOS versions older than Catalina
- */
-@available(*, deprecated, message: "the SDK will not support macOS versions older than Catalina")
-public func resumeAuth(_ urls: [URL]) {
-    guard let url = urls.first else { return }
-    _ = TransactionStore.shared.resume(url)
+    static func shared() -> NSApplication? {
+        return NSApplication.perform(NSSelectorFromString("sharedApplication"))?.takeUnretainedValue() as? NSApplication
+    }
+
 }
 
 extension ASTransaction: ASWebAuthenticationPresentationContextProviding {
