--- Auth0/NSApplication+Shared.swift
@@ -1,11 +1,0 @@
-#if os(macOS)
-import AppKit
-
-extension NSApplication {
-
-    static func shared() -> NSApplication? {
-        return NSApplication.perform(NSSelectorFromString("sharedApplication"))?.takeUnretainedValue() as? NSApplication
-    }
-
-}
-#endif
