--- Auth0/UIApplication+Shared.swift
@@ -1,11 +1,0 @@
-#if os(iOS)
-import UIKit
-
-extension UIApplication {
-
-    static func shared() -> UIApplication? {
-        return UIApplication.perform(NSSelectorFromString("sharedApplication"))?.takeUnretainedValue() as? UIApplication
-    }
-
-}
-#endif
